Refactored `configs.py` into separate scripts for clarity
